Refactor tubular context menus
