+++ Src/WitsmlExplorer.Frontend/components/ContextMenus/TubularObjectContextMenu.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import ContextMenu from "./ContextMenu";
 import { MenuItem } from "@material-ui/core";
 import OperationType from "../../contexts/operationType";
@@ -11,15 +11,12 @@ import { Server } from "../../models/server";
 import { Typography } from "@equinor/eds-core-react";
 import TubularReference from "../../models/jobs/tubularReference";
 import styled from "styled-components";
-import CredentialsService, { ServerCredentials } from "../../services/credentialsService";
-import { parseStringToTubularReference } from "../../models/jobs/copyTubularJob";
 import JobService, { JobType } from "../../services/jobService";
 import Wellbore from "../../models/wellbore";
-import WellboreReference from "../../models/jobs/wellboreReference";
-import UserCredentialsModal, { CredentialsMode, UserCredentialsModalProps } from "../Modals/UserCredentialsModal";
 import TubularService from "../../services/tubularService";
 import { UpdateWellboreTubularAction } from "../../contexts/navigationStateReducer";
 import ModificationType from "../../contexts/modificationType";
+import { onClickPaste, useClipboardTubularReference } from "./TubularContextMenuUtils";
 
 export interface TubularObjectContextMenuProps {
   dispatchNavigation: (action: UpdateWellboreTubularAction) => void;
@@ -32,20 +29,7 @@ export interface TubularObjectContextMenuProps {
 
 const TubularObjectContextMenu = (props: TubularObjectContextMenuProps): React.ReactElement => {
   const { dispatchNavigation, dispatchOperation, tubular, selectedServer, wellbore, servers } = props;
-  const [tubularReference, setTubularReference] = useState<TubularReference>(null);
-
-  useEffect(() => {
-    const tryToParseClipboardContent = async () => {
-      try {
-        const clipboardText = await navigator.clipboard.readText();
-        const tubularReference = parseStringToTubularReference(clipboardText);
-        setTubularReference(tubularReference);
-      } catch (e) {
-        //Not a valid object on the clipboard? That is fine, we won't use it.
-      }
-    };
-    tryToParseClipboardContent();
-  }, []);
+  const [tubularReference, setTubularReference] = useClipboardTubularReference();
 
   const deleteTubular = async () => {
     dispatchOperation({ type: OperationType.HideModal });
@@ -81,52 +65,6 @@ const TubularObjectContextMenu = (props: TubularObjectContextMenuProps): React.R
     dispatchOperation({ type: OperationType.HideContextMenu });
   };
 
-  const showCredentialsModal = (server: Server, errorMessage: string) => {
-    const onConnectionVerified = async (credentials: ServerCredentials) => {
-      await CredentialsService.saveCredentials(credentials);
-      orderCopyJob();
-      dispatchOperation({ type: OperationType.HideModal });
-    };
-
-    const currentCredentials = CredentialsService.getSourceServerCredentials();
-    const userCredentialsModalProps: UserCredentialsModalProps = {
-      server: server,
-      serverCredentials: currentCredentials,
-      mode: CredentialsMode.TEST,
-      errorMessage,
-      onConnectionVerified,
-      confirmText: "Save"
-    };
-    dispatchOperation({ type: OperationType.DisplayModal, payload: <UserCredentialsModal {...userCredentialsModalProps} /> });
-  };
-
-  const onClickPaste = async () => {
-    const sourceServer = servers.find((server) => server.url === tubularReference.serverUrl);
-    if (sourceServer !== null) {
-      CredentialsService.setSourceServer(sourceServer);
-      const hasPassword = CredentialsService.hasPasswordForServer(sourceServer);
-      if (!hasPassword) {
-        showCredentialsModal(
-          sourceServer,
-          `You are trying to paste a tubular from a server that you are not logged in to. Please provide username and password for ${sourceServer.name}.`
-        );
-      } else {
-        orderCopyJob();
-      }
-    }
-  };
-
-  const orderCopyJob = () => {
-    const wellboreReference: WellboreReference = {
-      wellUid: wellbore.wellUid,
-      wellboreUid: wellbore.uid
-    };
-
-    const copyJob = { source: tubularReference, target: wellboreReference };
-    JobService.orderJob(JobType.CopyTubular, copyJob);
-    dispatchOperation({ type: OperationType.HideContextMenu });
-  };
-
   const onClickDelete = async () => {
     const confirmation = (
       <ConfirmModal
@@ -151,7 +89,7 @@ const TubularObjectContextMenu = (props: TubularObjectContextMenuProps): React.R
           <StyledIcon name="copy" color={colors.interactive.primaryResting} />
           <Typography color={"primary"}>Copy tubular</Typography>
         </MenuItem>,
-        <MenuItem key={"paste"} onClick={() => onClickPaste()} disabled={tubularReference === null}>
+        <MenuItem key={"paste"} onClick={() => onClickPaste(servers, dispatchOperation, wellbore, tubularReference)} disabled={tubularReference === null}>
           <StyledIcon name="paste" color={colors.interactive.primaryResting} />
           <Typography color={"primary"}>Paste tubular</Typography>
         </MenuItem>,
