+++ Src/WitsmlExplorer.Frontend/components/ContextMenus/TubularsContextMenu.tsx
@@ -1,20 +1,14 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import ContextMenu from "./ContextMenu";
 import { MenuItem } from "@material-ui/core";
-import OperationType from "../../contexts/operationType";
 import Icon from "../../styles/Icons";
 import { colors } from "../../styles/Colors";
 import { DisplayModalAction, HideContextMenuAction, HideModalAction } from "../../contexts/operationStateReducer";
 import { Server } from "../../models/server";
 import { Typography } from "@equinor/eds-core-react";
-import TubularReference from "../../models/jobs/tubularReference";
 import styled from "styled-components";
-import CredentialsService, { ServerCredentials } from "../../services/credentialsService";
-import { parseStringToTubularReference } from "../../models/jobs/copyTubularJob";
-import JobService, { JobType } from "../../services/jobService";
 import Wellbore from "../../models/wellbore";
-import WellboreReference from "../../models/jobs/wellboreReference";
-import UserCredentialsModal, { CredentialsMode, UserCredentialsModalProps } from "../Modals/UserCredentialsModal";
+import { onClickPaste, useClipboardTubularReference } from "./TubularContextMenuUtils";
 
 export interface TubularsContextMenuProps {
   dispatchOperation: (action: HideModalAction | HideContextMenuAction | DisplayModalAction) => void;
@@ -24,71 +18,12 @@ export interface TubularsContextMenuProps {
 
 const TubularsContextMenu = (props: TubularsContextMenuProps): React.ReactElement => {
   const { dispatchOperation, wellbore, servers } = props;
-  const [tubularReference, setTubularReference] = useState<TubularReference>(null);
-
-  useEffect(() => {
-    const tryToParseClipboardContent = async () => {
-      try {
-        const clipboardText = await navigator.clipboard.readText();
-        const tubularReference = parseStringToTubularReference(clipboardText);
-        setTubularReference(tubularReference);
-      } catch (e) {
-        //Not a valid object on the clipboard? That is fine, we won't use it.
-      }
-    };
-    tryToParseClipboardContent();
-  }, []);
-
-  const showCredentialsModal = (server: Server, errorMessage: string) => {
-    const onConnectionVerified = async (credentials: ServerCredentials) => {
-      await CredentialsService.saveCredentials(credentials);
-      orderCopyJob();
-      dispatchOperation({ type: OperationType.HideModal });
-    };
-
-    const currentCredentials = CredentialsService.getSourceServerCredentials();
-    const userCredentialsModalProps: UserCredentialsModalProps = {
-      server: server,
-      serverCredentials: currentCredentials,
-      mode: CredentialsMode.TEST,
-      errorMessage,
-      onConnectionVerified,
-      confirmText: "Save"
-    };
-    dispatchOperation({ type: OperationType.DisplayModal, payload: <UserCredentialsModal {...userCredentialsModalProps} /> });
-  };
-
-  const onClickPaste = async () => {
-    const sourceServer = servers.find((server) => server.url === tubularReference.serverUrl);
-    if (sourceServer !== null) {
-      CredentialsService.setSourceServer(sourceServer);
-      const hasPassword = CredentialsService.hasPasswordForServer(sourceServer);
-      if (!hasPassword) {
-        showCredentialsModal(
-          sourceServer,
-          `You are trying to paste a tubular from a server that you are not logged in to. Please provide username and password for ${sourceServer.name}.`
-        );
-      } else {
-        orderCopyJob();
-      }
-    }
-  };
-
-  const orderCopyJob = () => {
-    const wellboreReference: WellboreReference = {
-      wellUid: wellbore.wellUid,
-      wellboreUid: wellbore.uid
-    };
-
-    const copyJob = { source: tubularReference, target: wellboreReference };
-    JobService.orderJob(JobType.CopyTubular, copyJob);
-    dispatchOperation({ type: OperationType.HideContextMenu });
-  };
+  const [tubularReference] = useClipboardTubularReference();
 
   return (
     <ContextMenu
       menuItems={[
-        <MenuItem key={"paste"} onClick={() => onClickPaste()} disabled={tubularReference === null}>
+        <MenuItem key={"paste"} onClick={() => onClickPaste(servers, dispatchOperation, wellbore, tubularReference)} disabled={tubularReference === null}>
           <StyledIcon name="paste" color={colors.interactive.primaryResting} />
           <Typography color={"primary"}>Paste tubular</Typography>
         </MenuItem>
